backend/requirements.txt
# BinBaz RAG System — Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# LLM
pydantic-ai-slim[groq]>=0.1.0

# Vector DB
qdrant-client>=1.12.0

# Embeddings
sentence-transformers>=3.0.0
fastembed>=0.4.0

# Data
orjson>=3.9.0
aiofiles>=24.0.0

# HTTP
httpx>=0.27.0
aiohttp>=3.10.0

# Monitoring
python-dotenv>=1.0.0
